More cleanup
After this set of changes, and some modifications to the `go.mod` the tests and code now compiles and runs. The things I've removed I don't believe are needed for this layer but feel free to comment on them.

My next step is to setup the CI so that even if failing the tests run properly and start a more aggressive refactor from there.