Metrics: Receivers and Integrations
Introduce a set of metrics to observe the integrations and recivers of an instance.

For the sake of keeping cardinality as low as possible receviers are group into two categories: active or inactive
However, for integrations I think it's useful to know how many we have of each type. I'd like to further know if an integration is active or not but decided not to do that know and roll with the current set in the meantime.