Add support for template testing (#77)
This method will use the given alerts as template context data and attempt to interpolate each of the top-level definition blocks passed in through the template field. A top-level definition is one that isn't used by any others.

Default templates and any already saved templates defined in the current AM as associated templates will be available for reference by the ones being tested. The exception to this is the saved template with the same name as the one passed via the name field. This is so that users can real-time test out modifications to existing templates (which is, ultimately, the end-goal of this endpoint).