Update Alertmanager fork
This commit updates our fork of Alertmanager to commit 323393f,
which is based on commit 1da134a in Prometheus Alertmanager.
It removes the overrides for validation of alerts, labels and
silences as UTF-8 is now supported upstream. The compat package,
which controls the enablement of UTF-8, is initialized in fallback
mode using the noop flags.