templates: add AlertingListURL to ExternalData
Some of the receivers use the /alerting/list for grouping alerts
when their generator URLs aren't the same.

However these alerting list URLs are not "org-aware".

Instead of passing an orgId to every receiver constructor,
we can enrich the template (extended) data with a URL that
has the ?orgID=N in case it exists in the common annotations,
to be reused by any receiver.