Simplify image Provider interface
Previous abstraction change introduced GetImageURL and GetRawImage.
These methods were meant to replace GetImage, however only Discord was
using it.

Switching to these methods would require significant rewrites in
integrations to not use WithStoredImages. So, instead I changed the
abstraction to one that is more compatible with WithStoredImages.

This new abstraction satisfies the same requirements as the previous
in that it hides Path and File information as implementation details
so that Providers can control access to raw data.